size_t should be 64-bits on a 64-bit machine. mmsize_t should be 32-bits (unless CONFIG_MM_SMALL is selected).  This commit backs out the last change to stddef.h

+++ include/nuttx/mm/mm.h
@@ -186,7 +186,7 @@
    typedef uint16_t mmsize_t;
 #  define MMSIZE_MAX 0xffff
 #else
-   typedef size_t mmsize_t;
+   typedef uint32_t mmsize_t;
 #  define MMSIZE_MAX SIZE_MAX
 #endif
